wip services, rework models

+++ backend/src/main/java/com/example/backend/repository/QuantityProductRepository.java
@@ -3,5 +3,5 @@ package com.example.backend.repository;
 import com.example.backend.model.QuantityProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuantityProductRepository extends JpaRepository<QuantityProduct, Long> {
+public interface QuantityProductRepository extends JpaRepository<QuantityProduct, Integer> {
 }
